requirements.txt
fastapi~=0.103.1
numpy~=1.22.3
pandas~=1.4.2
python-dotenv~=1.0.0
